test: consolidate clone detector benchmarks
